Пропускная система/requirements.txt
requests==2.28.2
python-dotenv==1.0.0
PySimpleGUI==4.60.4